# src/openapi.yaml
openapi: 3.0.3
info:
  title: Docker Admin API
  version: '1.0.0'
  description: Protocol-compliant API for Docker resource management
servers:
  - url: /api
paths:
...
  /execs:
    get:
      summary: List execs
      responses:
        '200':
          description: List of execs
    post:
      summary: Create a global exec
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Exec created
  /execs/{id}:
    delete:
      summary: Remove an exec
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Exec removed
...